[Rename] 전체 관리자 관련 파일 > 어드민 폴더로 이동
